feat(tool-stack): move proxy command into tool
