--- packages/@code-like-a-carpenter/cli-plugin-proxy/README.md
@@ -1,34 +1,0 @@
-# @code-like-a-carpenter/cli-plugin-proxy
-
-[![standard-readme compliant](https://img.shields.io/badge/readme%20style-standard-brightgreen.svg?style=flat-square)](https://github.com/RichardLitt/standard-readme)
-
-> Proxy command which forwards one or more API Gateway endpoints to localhost
-
-## Table of Contents
-
--   [Install](#install)
--   [Usage](#usage)
--   [Maintainer](#maintainer)
--   [Contributing](#contributing)
--   [License](#license)
-
-## Install
-
-```bash
-npm i @code-like-a-carpenter/cli-plugin-proxy
-```
-
-## Usage
-
-## Maintainer
-
-[Ian Remmel](https://www.ianwremmel.com)
-
-## Contributing
-
-Please see contributing guidelines at the
-[project homepage](https://www.github.com/code-like-a-carpenter/workbench/).
-
-## License
-
-MIT © [Ian Remmel](https://www.ianwremmel.com) 2023 until at least now
