--- packages/@code-like-a-carpenter/cli-plugin-proxy/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@code-like-a-carpenter/cli-plugin-proxy",
-  "description": "Proxy command which forwards one or more API Gateway endpoints to localhost",
-  "dependencies": {
-    "@aws-sdk/client-cloudformation": "3.188.0",
-    "@code-like-a-carpenter/assert": "*",
-    "@code-like-a-carpenter/cli-core": "*",
-    "@code-like-a-carpenter/tooling-common": "*",
-    "express": "^4.18.2",
-    "http-proxy": "^1.18.1",
-    "lodash.snakecase": "^4.1.1",
-    "vhost": "^3.0.2"
-  },
-  "author": "Ian Remmel <1182361+ianwremmel@users.noreply.github.com> (https://www.ianwremmel.com)",
-  "bugs": "https://www.github.com/code-like-a-carpenter/workbench/issues",
-  "engines": {
-    "node": "20.x"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/types/index.d.ts",
-      "development": "./src/index.ts",
-      "import": "./dist/esm/index.mjs"
-    },
-    "./package.json": "./package.json"
-  },
-  "homepage": "https://www.github.com/code-like-a-carpenter/workbench/tree/main/packages/@code-like-a-carpenter/cli-plugin-proxy",
-  "license": "MIT",
-  "repository": "git@github.com:code-like-a-carpenter/workbench.git",
-  "types": "dist/types",
-  "publishConfig": {
-    "access": "public"
-  },
-  "bin": "./cli.mjs",
-  "devDependencies": {
-    "@types/http-proxy": "^1.17.14",
-    "@types/lodash.snakecase": "^4.1.9",
-    "@types/vhost": "^3.0.9"
-  },
-  "releaseAll": 1
-}
