--- packages/@code-like-a-carpenter/cli-plugin-proxy/src/index.ts
@@ -1,2 +1,0 @@
-export {plugin as default} from './plugin';
-export * from './proxy';
