--- packages/@code-like-a-carpenter/cli-plugin-proxy/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "outDir": "./dist/types",
-    "rootDir": "./src"
-  },
-  "extends": "../../../tsconfig.references.json",
-  "include": ["src"],
-  "references": [
-    {
-      "path": "../assert"
-    },
-    {
-      "path": "../cli-core"
-    },
-    {
-      "path": "../tooling-common"
-    }
-  ]
-}
