+++ packages/@code-like-a-carpenter/cli/tsconfig.json
@@ -13,9 +13,6 @@
       "path": "../cli-plugin-example"
     },
     {
-      "path": "../cli-plugin-proxy"
-    },
-    {
       "path": "../tool-stack"
     },
     {
